removed driver argument as it isn't necessary, all drivers are just tested and run against to find one that works.
there is a small note to make about the use of pvl.loads() in this script. I commented out 2 lines using that function and had to change to variable names in the subsequent lines as this pvl.loads() function won't work as expected until #587 is merged to apply the bug fix to that method; that fix will allow ale.loads() to read in PVLObjects, currently because the lines are commented ale.loads() just reads in the image files themselves which still works but requires more overhead in ale.loads().