added a test for pvl.load/ale.loads on PVLObjects
Not completely done as I need to figure out how to load some kind of test cube akin to the one I am using inside of the test itself without loading it from local storage. There should be some way to put it in the test data area my main problem is I would like to avoid putting a whole cube in the test's data folder as they are larger files and I believe that convention of putting cubes there is to be avoided.