Update to isd_generate.py to allow for a NAIF radius override (#486)
* Update to isd_generate.py to allow for a NAIF radius override

Based on this RFC, we want to change our Mars pipeline to use a sphere. https://astrodiscuss.usgs.gov/t/appl-rfc-use-iau-sphere-for-hirise-dtm-projects-in-socet-set/428

Currently ALE uses the Mars elliptical defaults from the NAIF kernels. This update allows the user to override the kernels and set a spherical radius for both semimajor and semiminor in the ISD.

This is not only needed for Mars. For triaxial bodies, the IAU recommends a best-fit sphere should also be set for map products. References (DOI) are listed in the updated help for the current IAU values.

* remove extra file write if new radius is sent

* update to allow ellipse, minor help update

Add --semimajor and --semiminor parameters. I tried nargs="+" for --radius but fought for too long with argparse to do all the checking (like checking it is float). I am fine with these options but happy to hear better ways to use argparse.

* Update isd_generate.py (#619)

add in PROJ like options for radius names ("-a", "-r", "--radius", "-b")

* Radius update with changes to address original PR and review comments (#620)

* Update isd_generate.py

add in PROJ like options for radius names ("-a", "-r", "--radius", "-b")

* Update isd_generate.py

updated required statement for --semimajor to catch if only --semiminor or -b is sent. Request that radius is sent in meters and updated some typos.

---------

Co-authored-by: Austin Sanders <arsanders@usgs.gov>

---------

Co-authored-by: Austin Sanders <arsanders@usgs.gov>